+++ NEWS.md
@@ -1,7 +1,7 @@
 # Rcompadre 1.3.0
 
 * Minor performance enhancements, some of which speed up functions while others improve code readability and stability.
-* Added functions `mpm_sd`/`mat_sd` and `mpm_median`/`mat_median` which join the existing `mpm_mean`/`mat_mean` functions to calculate element-wise summary statistics across MPMs of the same dimension. Code contributed by Darren Norris with modifications by Owen Jones.
+* Added functions `mpm_sd`/`mat_sd` and `mpm_median`/`mat_median` which join the existing `mpm_mean`/`mat_mean` functions to calculate element-wise summary statistics across MPMs of the same dimension. Code contributed by Darren Norris with modifications by Owen Jones. A future version will include a generic "summarise matrix" function.
 * Increased test coverage.
 
 # Rcompadre 1.2.1